Moved DurationGrid to its proper folder
