+++ src/components/settings/controls/DurationGrid/Note.tsx
