+++ src/components/settings/controls/DurationGrid/durationGridStyles.ts
