+++ src/components/settings/pages/PageRests.tsx
@@ -2,7 +2,7 @@ import { FC } from 'react'
 import { useTranslation } from 'react-i18next'
 import { Section, Description, Label } from '../controls/InputSectionPrimitives'
 import { SettingsPageProps } from '../types'
-import { DurationGrid } from '../controls/DurationGrid2/DurationGrid'
+import { DurationGrid } from '../controls/DurationGrid/DurationGrid'
 import { DurationConfig } from '../../../state/types'
 import { DurationType } from '../../../common/durationType'
